// src/web/shared/proto/jaiabot/messages/mission.ts
export interface Speeds {
    transit?: number | undefined;
    stationkeep_outer?: number | undefined;
}

export interface BottomDepthSafetyParams {
    constant_heading?: number | undefined;
    constant_heading_time?: number | undefined;
    constant_heading_speed?: number | undefined;
    safety_depth?: number | undefined;
}
// ...
export interface MissionTask {
    type?: MissionTask_TaskType | undefined;
    /** for type == DIVE */
    dive?: MissionTask_DiveParameters | undefined;
    /** for type == SURFACE_DRIFT and type == DIVE */
    surface_drift?: MissionTask_DriftParameters | undefined;
    /** for type == CONSTANT_HEADING */
    constant_heading?: MissionTask_ConstantHeadingParameters | undefined;
    start_pam?: boolean | undefined;
    /** for type == STATION_KEEP */
    station_keep?: MissionTask_StationKeepParameters | undefined;
}

export enum MissionTask_TaskType {
    NONE = "NONE",
    DIVE = "DIVE",
    STATION_KEEP = "STATION_KEEP",
    SURFACE_DRIFT = "SURFACE_DRIFT",
    CONSTANT_HEADING = "CONSTANT_HEADING",
}

export interface MissionTask_DiveParameters {
    max_depth?: number | undefined;
    depth_interval?: number | undefined;
    hold_time?: number | undefined;
    bottom_dive?: boolean | undefined;
}

export interface MissionTask_DriftParameters {
    drift_time?: number | undefined;
}

export interface MissionTask_ConstantHeadingParameters {
    constant_heading?: number | undefined;
    constant_heading_time?: number | undefined;
    constant_heading_speed?: number | undefined;
}

export interface MissionTask_StationKeepParameters {
    station_keep_time?: number | undefined;
}

export interface MissionPlan {
    start?: MissionPlan_MissionStart | undefined;
    movement?: MissionPlan_MovementType | undefined;
    goal?: MissionPlan_Goal[] | undefined;
    recovery?: MissionPlan_Recovery | undefined;
    speeds?: Speeds | undefined;
    bottom_depth_safety_params?: BottomDepthSafetyParams | undefined;
    fragment_index?: number | undefined;
    expected_fragments?: number | undefined;
    trail?: MissionPlan_TrailParam | undefined;
    repeats?: number | undefined;
    /** not sent over radio, only saved to the hub */
    mission_name?: string | undefined;
}

export enum MissionPlan_MissionStart {
    START_IMMEDIATELY = "START_IMMEDIATELY",
    /**
     * START_ON_COMMAND - START_IN_WATER_DETECT = 3;
     * START_AFTER_DELAY = 4;
     */
    START_ON_COMMAND = "START_ON_COMMAND",
}

export enum MissionPlan_MovementType {
    TRANSIT = "TRANSIT",
    REMOTE_CONTROL = "REMOTE_CONTROL",
    TRAIL = "TRAIL",
}

export interface MissionPlan_Goal {
    name?: string | undefined;
    location?: GeographicCoordinate | undefined;
    task?: MissionTask | undefined;
    moveWptMode?: boolean | undefined;
}

export interface MissionPlan_Recovery {
    recover_at_final_goal?: boolean | undefined;
    location?: GeographicCoordinate | undefined;
}

export interface MissionPlan_TrailParam {
    contact?: number | undefined;
    angle?: number | undefined;
    angle_relative?: boolean | undefined;
    range?: number | undefined;
}